리팩토링중...

+++ src/main/java/kr/hhplus/be/server/domain/user/adapter/in/web/request/BalanceChargeRequest.java
@@ -1,10 +1,10 @@
 package kr.hhplus.be.server.domain.user.adapter.in.web.request;
 
+import java.math.BigDecimal;
+
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
-
-import java.math.BigInteger;
 
 @Getter
 @Schema(name = "BalanceChargeRequest", description = "잔액 충전 Request")
@@ -12,5 +12,5 @@ public class BalanceChargeRequest {
 
     @Schema(description = "충전 잔액")
     @NotNull
-    private BigInteger amount;
+    private BigDecimal amount;
 }
