+++ src/main/java/kr/hhplus/be/server/domain/user/adapter/in/web/response/BalanceResponse.java
@@ -1,10 +1,10 @@
 package kr.hhplus.be.server.domain.user.adapter.in.web.response;
 
+import java.math.BigDecimal;
+
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 import lombok.Getter;
-
-import java.math.BigInteger;
 
 @Getter
 @Builder
@@ -12,5 +12,5 @@ import java.math.BigInteger;
 public class BalanceResponse {
 
     @Schema(description = "잔액")
-    private BigInteger amount;
+    private BigDecimal amount;
 }
